moved jest config to package.json; moved tests into src directory
