--- source/nodejs/adaptivecards-fabric/jest.config.js
@@ -1,12 +1,0 @@
-module.exports = {
-    preset: 'ts-jest/presets/js-with-ts',
-    verbose: true,
-    moduleFileExtensions: ["ts", "tsx", "js", "jsx", "json", "node"],
-    // testRegex: "(/__tests__/.*|(\\.|/)(test|spec))\\.(jsx?|tsx?)$",
-    testPathIgnorePatterns: [
-        "/node_modules/"
-    ],
-    modulePathIgnorePatterns: [
-        "TestUtils"
-    ]
-};
