+++ source/nodejs/adaptivecards-fabric/src/__tests__/Components/Inputs/InputTextFabric.test.tsx
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { InputTextFabric } from "../../../src/Components/Inputs/InputTextFabric";
+import { InputTextFabric } from "../../../Components/Inputs/InputTextFabric";
 import * as TestUtils from "../../TestUtils";
 import { initializeIcons } from "office-ui-fabric-react";
