+++ source/nodejs/adaptivecards-fabric/src/__tests__/Components/Inputs/InputTimeFabric.test.tsx
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { InputTimeFabric } from "../../../src/Components/Inputs/InputTimeFabric";
+import { InputTimeFabric } from "../../../Components/Inputs/InputTimeFabric";
 import * as TestUtils from "../../TestUtils";
 import { initializeIcons } from "office-ui-fabric-react";
