+++ source/nodejs/adaptivecards-fabric/tsconfig.json
@@ -26,7 +26,7 @@
 	},
 	"include": ["src"],
 	"exclude": [
-		"__tests__",
+		"src/__tests__",
 		"node_modules"
 	]
 }
